Fix adding items to the inventory
Fixed adding items to the inventory. Mainly a fix for sponge with bukkit support pending. There were multiple issues that made this a problem, the scheduler workaround was the initial problem. This described problem is solved by setting the item first and then running the scheduler after to set it again if it has not been done already.