Gui title animation test
Initial test of gui title animation. There are a few hacks that seem to make this more stable. Those being delaying the title update by one tick and checking if the gui is open and checking if the gui is open on inventory close. Seems to be stable at 10 ticks and semi-stable at 5 ticks.