Handle drag event, add SetMoveable function
Now handle the drag event, this was never handled in DynamicGui so this could be the cause of some players losing their items. Stacks apparently are dragged even if the stack size is 1. The drag event doesn't do much, currently it just checks if an item should be able to be put into a slot. If an item shouldn't be in a slot the whole event is cancelled. The SetMoveable function is added to start testing functionality of moveable slots before a parser update.